arc: Add support to cache extracted raw videos in encode tests.
As the raw video files used to test encoding are too large to download
and store on a DUT, videos are stored in the vp9.webm format. Before
each test these videos are extracted to create the raw input video.
This process can take a lot of time, so this CL adds support to cache
the extracted raw video file. This saves about 30s test time when
running 1080p encode testcases. The feature is disabled by default as
we don't want to have DUT's run out of memory in the lab.

BUG=None
TEST=various video encode tests on Hatch

+++ src/chromiumos/tast/local/bundles/cros/arc/power_video_encode_perf.go
@@ -69,6 +69,8 @@ func init() {
 
 func PowerVideoEncodePerf(ctx context.Context, s *testing.State) {
 	const (
+		// Enable to cache the extracted raw video to speed up the test.
+		cacheExtractedVideo = true
 		// arcFilePath must be on the sdcard because of android permissions
 		arcFilePath = "/sdcard/Download/c2_e2e_test/"
 		logFileName = "gtest_logs.txt"
@@ -94,11 +96,20 @@ func PowerVideoEncodePerf(ctx context.Context, s *testing.State) {
 	}
 
 	// Extract video to create the raw video stream that will be encoded.
-	rawVideoPath, err := encoding.PrepareYUV(ctx, s.DataPath(opts.Params.Name), opts.PixelFormat, opts.Params.Size)
+	rawVideoPath := strings.TrimSuffix(s.DataPath(opts.Params.Name), ".vp9.webm")
+	switch opts.PixelFormat {
+	case videotype.I420:
+		rawVideoPath += ".i420.yuv"
+	case videotype.NV12:
+		rawVideoPath += ".nv12.yuv"
+	}
+	err = encoding.PrepareYUV(ctx, s.DataPath(opts.Params.Name), rawVideoPath, opts.PixelFormat, opts.Params.Size)
 	if err != nil {
 		s.Fatal("Failed to prepare YUV file: ", err)
 	}
-	defer os.Remove(rawVideoPath)
+	if !cacheExtractedVideo {
+		defer os.Remove(rawVideoPath)
+	}
 
 	// Set up device for measuring power drain.
 	sup, cleanup := setup.New("video power")
